implement processing synchronized and opened action

+++ Source/Application/Kysect.Shreks.Application.Abstractions/Submissions/Commands/CreateOrUpdateSubmission.cs
@@ -5,7 +5,7 @@ namespace Kysect.Shreks.Application.Abstractions.Submissions.Commands;
 
 public static class CreateOrUpdateSubmissionCommand
 {
-    public record Command(Guid StudentId, Guid AssignmentId, string Payload, string Organization, string Repository, int PrNumber) : IRequest<Response>;
+    public record Command(Guid StudentId, Guid AssignmentId, string Payload, string Organization, string Repository, long PrNumber) : IRequest<Response>;
 
     public record Response(SubmissionDto Submission);
 }
